chore: move generateTopWrapper
