+++ FABulous/fabric_generator/fabric_gen.py
@@ -16,7 +16,6 @@
 
 
 import os
-import re
 import string
 from pathlib import Path
 from typing import TYPE_CHECKING
@@ -24,12 +23,7 @@ from typing import TYPE_CHECKING
 from loguru import logger
 
 from FABulous.custom_exception import InvalidFileType
-from FABulous.fabric_definition.define import (
-    IO,
-)
 from FABulous.fabric_definition.Fabric import Fabric
-from FABulous.fabric_generator.code_generation_Verilog import VerilogWriter
-from FABulous.fabric_generator.code_generation_VHDL import VHDLWriter
 from FABulous.fabric_generator.code_generator import codeGenerator
 from FABulous.fabric_generator.parser.parse_configmem import parseConfigMem
 from FABulous.fabric_generator.parser.parse_hdl import parseBelFile
@@ -68,326 +62,6 @@ class FabricGenerator:
         logger.info(
             f"Generate switch matrix debug signals: {self.switch_matrix_debug_signal}"
         )
-
-    def generateTopWrapper(self) -> None:
-        """Generate the top wrapper of the fabric including features that are not
-        located inside the fabric such as BRAM."""
-
-        def split_port(p):
-            # split a port according to how we want to sort external ports:
-            # ((y, x), (indices...), basename)
-            # Tile_X9Y6_RAM2FAB_D1_I0 --> ((6, 9), (1, 0), "RAM2FAB_D_I")
-            m = re.match(r"Tile_X(\d+)Y(\d+)_(.*)", p)
-            x = int(m.group(1))
-            y = int(m.group(2))
-            port = m.group(3)
-
-            basename = ""
-            numbuf = ""
-            indices = []
-            got_split = False
-            for ch in port:
-                if ch.isnumeric() and got_split:
-                    numbuf += ch
-                else:
-                    if ch == "_":
-                        # this way we treat the 2 in RAM2FAB as part of the name, rather than an index
-                        got_split = True
-                    if numbuf != "":
-                        indices.append(int(numbuf))
-                    basename += ch
-
-            if numbuf != "":
-                indices.append(int(numbuf))
-
-            # some backwards compat
-            basename = basename.removesuffix("_bit")
-            # top level IO has A and B parts combined and reverse order
-            if len(basename) == 7 and basename[1:] in ("_I_top", "_O_top", "_T_top"):
-                assert basename[0] in "ABCDEFGH"
-                indices.append(-(ord(basename[0]) - ord("A")))
-                basename = basename[2:]
-
-            # Y is in reverse order
-            return ((-y, x), tuple(indices), basename)
-
-        # determine external ports so we can group them
-        externalPorts = []
-        portGroups = dict()
-        for y, row in enumerate(self.fabric.tile):
-            for x, tile in enumerate(row):
-                if tile is not None:
-                    for bel in tile.bels:
-                        for i in bel.externalInput:
-                            externalPorts.append((IO.INPUT, f"Tile_X{x}Y{y}_{i}"))
-                        for i in bel.externalOutput:
-                            externalPorts.append((IO.OUTPUT, f"Tile_X{x}Y{y}_{i}"))
-        for iodir, name in externalPorts:
-            _yx, _indices, port = split_port(name)
-            if port not in portGroups:
-                portGroups[port] = (iodir, [])
-            portGroups[port][1].append(name)
-        # sort port groups according to vectorisation order
-        for _name, g in portGroups.items():
-            g[1].sort(key=lambda x: split_port(x))
-
-        # header
-        numberOfRows = self.fabric.numberOfRows - 2
-        numberOfColumns = self.fabric.numberOfColumns
-        self.writer.addHeader(f"{self.fabric.name}_top")
-        self.writer.addParameterStart(indentLevel=1)
-        self.writer.addParameter("include_eFPGA", "integer", 1, indentLevel=2)
-        self.writer.addParameter("NumberOfRows", "integer", numberOfRows, indentLevel=2)
-        self.writer.addParameter(
-            "NumberOfCols", "integer", self.fabric.numberOfColumns, indentLevel=2
-        )
-        self.writer.addParameter(
-            "FrameBitsPerRow", "integer", self.fabric.frameBitsPerRow, indentLevel=2
-        )
-        self.writer.addParameter(
-            "MaxFramesPerCol", "integer", self.fabric.maxFramesPerCol, indentLevel=2
-        )
-        self.writer.addParameter(
-            "desync_flag", "integer", self.fabric.desync_flag, indentLevel=2
-        )
-        self.writer.addParameter(
-            "FrameSelectWidth", "integer", self.fabric.frameSelectWidth, indentLevel=2
-        )
-        self.writer.addParameter(
-            "RowSelectWidth", "integer", self.fabric.rowSelectWidth, indentLevel=2
-        )
-        self.writer.addParameterEnd(indentLevel=1)
-        self.writer.addPortStart(indentLevel=1)
-
-        self.writer.addComment("External IO port", onNewLine=True, indentLevel=2)
-        for name, group in sorted(portGroups.items(), key=lambda x: x[0]):
-            if self.fabric.numberOfBRAMs > 0 and (
-                "RAM2FAB" in name or "FAB2RAM" in name
-            ):
-                continue
-            self.writer.addPortVector(name, group[0], len(group[1]) - 1, indentLevel=2)
-        self.writer.addComment("Config related ports", onNewLine=True, indentLevel=2)
-        self.writer.addPortScalar("CLK", IO.INPUT, indentLevel=2)
-        self.writer.addPortScalar("resetn", IO.INPUT, indentLevel=2)
-        self.writer.addPortScalar("SelfWriteStrobe", IO.INPUT, indentLevel=2)
-        self.writer.addPortVector(
-            "SelfWriteData", IO.INPUT, self.fabric.frameBitsPerRow - 1, indentLevel=2
-        )
-        self.writer.addPortScalar("Rx", IO.INPUT, indentLevel=2)
-        self.writer.addPortScalar("ComActive", IO.OUTPUT, indentLevel=2)
-        self.writer.addPortScalar("ReceiveLED", IO.OUTPUT, indentLevel=2)
-        self.writer.addPortScalar("s_clk", IO.INPUT, indentLevel=2)
-        self.writer.addPortScalar("s_data", IO.INPUT, indentLevel=2)
-        self.writer.addPortEnd()
-        self.writer.addHeaderEnd(f"{self.fabric.name}_top")
-        self.writer.addDesignDescriptionStart(f"{self.fabric.name}_top")
-
-        # all the wires/connection with in the design
-        if "RAM2FAB_D_I" in portGroups and self.fabric.numberOfBRAMs > 0:
-            self.writer.addComment("BlockRAM ports", onNewLine=True)
-            self.writer.addNewLine()
-            self.writer.addConnectionVector("RAM2FAB_D_I", f"{numberOfRows * 4 * 4}-1")
-            self.writer.addConnectionVector("FAB2RAM_D_O", f"{numberOfRows * 4 * 4}-1")
-            self.writer.addConnectionVector("FAB2RAM_A_O", f"{numberOfRows * 4 * 2}-1")
-            self.writer.addConnectionVector("FAB2RAM_C_O", f"{numberOfRows * 4}-1")
-
-        self.writer.addNewLine()
-        self.writer.addComment("Signal declarations", onNewLine=True)
-        self.writer.addConnectionVector(
-            "FrameRegister", "(NumberOfRows*FrameBitsPerRow)-1"
-        )
-        self.writer.addConnectionVector(
-            "FrameSelect", "(MaxFramesPerCol*NumberOfCols)-1"
-        )
-        self.writer.addConnectionVector(
-            "FrameData", "(FrameBitsPerRow*(NumberOfRows+2))-1"
-        )
-        self.writer.addConnectionVector("FrameAddressRegister", "FrameBitsPerRow-1")
-        self.writer.addConnectionScalar("LongFrameStrobe")
-        self.writer.addConnectionVector("LocalWriteData", 31)
-        self.writer.addConnectionScalar("LocalWriteStrobe")
-        self.writer.addConnectionVector("RowSelect", "RowSelectWidth-1")
-
-        if isinstance(self.writer, VHDLWriter):
-            basePath = Path(self.writer.outFileName).parent
-            if not os.path.exists(f"{basePath}/Frame_Data_Reg.vhdl"):
-                raise FileExistsError(
-                    "Frame_Data_Reg.vhdl not found in the 'Fabric' directory."
-                )
-            if not os.path.exists(f"{basePath}/Frame_Select.vhdl"):
-                raise FileExistsError(
-                    "Frame_Select.vhdl not found in the 'Fabric' directory."
-                )
-            if not os.path.exists(f"{basePath}/eFPGA_Config.vhdl"):
-                raise FileExistsError(
-                    "Config.vhdl not found in the 'Fabric' directory."
-                )
-            if not os.path.exists(f"{basePath}/eFPGA.vhdl"):
-                raise FileExistsError(
-                    "eFPGA.vhdl not found in the 'Fabric' directory, need to generate the eFPGA first."
-                )
-            if not os.path.exists(f"{basePath}/BlockRAM_1KB.vhdl"):
-                raise FileExistsError(
-                    "BlockRAM_1KB.vhdl not found in the 'Fabric' directory."
-                )
-            self.writer.addComponentDeclarationForFile(
-                f"{basePath}/Frame_Data_Reg.vhdl"
-            )
-            self.writer.addComponentDeclarationForFile(f"{basePath}/Frame_Select.vhdl")
-            self.writer.addComponentDeclarationForFile(f"{basePath}/eFPGA_Config.vhdl")
-            self.writer.addComponentDeclarationForFile(f"{basePath}/eFPGA.vhdl")
-            self.writer.addComponentDeclarationForFile(f"{basePath}/BlockRAM_1KB.vhdl")
-
-        self.writer.addLogicStart()
-
-        if isinstance(self.writer, VerilogWriter):
-            self.writer.addPreprocIfNotDef("EMULATION")
-
-        # the config module
-        self.writer.addNewLine()
-        self.writer.addInstantiation(
-            compName="eFPGA_Config",
-            compInsName="eFPGA_Config_inst",
-            portsPairs=[
-                ("CLK", "CLK"),
-                ("resetn", "resetn"),
-                ("Rx", "Rx"),
-                ("ComActive", "ComActive"),
-                ("ReceiveLED", "ReceiveLED"),
-                ("s_clk", "s_clk"),
-                ("s_data", "s_data"),
-                ("SelfWriteData", "SelfWriteData"),
-                ("SelfWriteStrobe", "SelfWriteStrobe"),
-                ("ConfigWriteData", "LocalWriteData"),
-                ("ConfigWriteStrobe", "LocalWriteStrobe"),
-                ("FrameAddressRegister", "FrameAddressRegister"),
-                ("LongFrameStrobe", "LongFrameStrobe"),
-                ("RowSelect", "RowSelect"),
-            ],
-            paramPairs=[
-                ("RowSelectWidth", "RowSelectWidth"),
-                ("NumberOfRows", "NumberOfRows"),
-                ("desync_flag", "desync_flag"),
-                ("FrameBitsPerRow", "FrameBitsPerRow"),
-            ],
-        )
-        self.writer.addNewLine()
-
-        # the frame data reg module
-        for row in range(numberOfRows):
-            self.writer.addInstantiation(
-                compName="Frame_Data_Reg",
-                compInsName=f"inst_Frame_Data_Reg_{row}",
-                portsPairs=[
-                    ("FrameData_I", "LocalWriteData"),
-                    (
-                        "FrameData_O",
-                        f"FrameRegister[{row}*FrameBitsPerRow+FrameBitsPerRow-1:{row}*FrameBitsPerRow]",
-                    ),
-                    ("RowSelect", "RowSelect"),
-                    ("CLK", "CLK"),
-                ],
-                paramPairs=[
-                    ("FrameBitsPerRow", "FrameBitsPerRow"),
-                    ("RowSelectWidth", "RowSelectWidth"),
-                    ("Row", str(row + 1)),
-                ],
-            )
-        self.writer.addNewLine()
-
-        # the frame select module
-        for col in range(numberOfColumns):
-            self.writer.addInstantiation(
-                compName="Frame_Select",
-                compInsName=f"inst_Frame_Select_{col}",
-                portsPairs=[
-                    ("FrameStrobe_I", "FrameAddressRegister[MaxFramesPerCol-1:0]"),
-                    (
-                        "FrameStrobe_O",
-                        f"FrameSelect[{col}*MaxFramesPerCol+MaxFramesPerCol-1:{col}*MaxFramesPerCol]",
-                    ),
-                    (
-                        "FrameSelect",
-                        "FrameAddressRegister[FrameBitsPerRow-1:FrameBitsPerRow-FrameSelectWidth]",
-                    ),
-                    ("FrameStrobe", "LongFrameStrobe"),
-                ],
-                paramPairs=[
-                    ("MaxFramesPerCol", "MaxFramesPerCol"),
-                    ("FrameSelectWidth", "FrameSelectWidth"),
-                    ("Col", str(col)),
-                ],
-            )
-        self.writer.addNewLine()
-
-        if isinstance(self.writer, VerilogWriter):
-            self.writer.addPreprocEndif()
-
-        # the fabric module
-        portList = []
-        signal = []
-
-        # external ports (IO, config access, BRAM, etc)
-        for name, group in sorted(portGroups.items(), key=lambda x: x[0]):
-            for i, sig in enumerate(group[1]):
-                portList.append(sig)
-                signal.append(f"{name}[{i}]")
-
-        portList.append("UserCLK")
-        signal.append("CLK")
-
-        portList.append("FrameData")
-        signal.append("FrameData")
-
-        portList.append("FrameStrobe")
-        signal.append("FrameSelect")
-
-        assert len(portList) == len(signal)
-        self.writer.addInstantiation(
-            compName=self.fabric.name,
-            compInsName=f"{self.fabric.name}_inst",
-            portsPairs=list(zip(portList, signal, strict=False)),
-        )
-
-        self.writer.addNewLine()
-
-        # the BRAM module
-        if "RAM2FAB_D_I" in portGroups and self.fabric.numberOfBRAMs > 0:
-            data_cap = int((numberOfRows * 4 * 4) / (self.fabric.numberOfBRAMs - 1))
-            addr_cap = int((numberOfRows * 4 * 2) / (self.fabric.numberOfBRAMs - 1))
-            config_cap = int((numberOfRows * 4) / (self.fabric.numberOfBRAMs - 1))
-            for i in range(self.fabric.numberOfBRAMs - 1):
-                portsPairs = [
-                    ("clk", "CLK"),
-                    ("rd_addr", f"FAB2RAM_A_O[{addr_cap * i + 8 - 1}:{addr_cap * i}]"),
-                    ("rd_data", f"RAM2FAB_D_I[{data_cap * i + 32 - 1}:{data_cap * i}]"),
-                    (
-                        "wr_addr",
-                        f"FAB2RAM_A_O[{addr_cap * i + 16 - 1}:{addr_cap * i + 8}]",
-                    ),
-                    ("wr_data", f"FAB2RAM_D_O[{data_cap * i + 32 - 1}:{data_cap * i}]"),
-                    ("C0", f"FAB2RAM_C_O[{config_cap * i}]"),
-                    ("C1", f"FAB2RAM_C_O[{config_cap * i + 1}]"),
-                    ("C2", f"FAB2RAM_C_O[{config_cap * i + 2}]"),
-                    ("C3", f"FAB2RAM_C_O[{config_cap * i + 3}]"),
-                    ("C4", f"FAB2RAM_C_O[{config_cap * i + 4}]"),
-                    ("C5", f"FAB2RAM_C_O[{config_cap * i + 5}]"),
-                ]
-                self.writer.addInstantiation(
-                    compName="BlockRAM_1KB",
-                    compInsName=f"Inst_BlockRAM_{i}",
-                    portsPairs=portsPairs,
-                )
-        if isinstance(self.writer, VHDLWriter):
-            self.writer.addAssignScalar(
-                "FrameData", ['X"12345678"', "FrameRegister", 'X"12345678"']
-            )
-        else:
-            self.writer.addAssignScalar(
-                "FrameData", ["32'h12345678", "FrameRegister", "32'h12345678"]
-            )
-        self.writer.addDesignDescriptionEnd()
-        self.writer.writeToFile()
 
     def generateBitsStreamSpec(self) -> dict[str, dict]:
         """Generate the bitstream specification of the fabric. This is needed and will
